Create modelREADME.md
Model's Role in the Research:
This model seeks to quantitatively assess the relationship between couples' financial organisation methods and their overall relationship satisfaction. The analysis focuses on correlating the textual descriptions of financial arrangements (mm_explain) with the self-reported levels of relationship happiness (part10). The outcomes of this model are anticipated to provide empirical support for potential legal reforms and offer a deeper understanding of the dynamics in non-marital cohabitations.